api/conda-requirements.txt
# Use this file to create a conda environment using:
# conda create -n <name> --file conda-requirements.txt

# Mandatory dependencies
flask
iris
python-dateutil
multipledispatch
boto3
